feat: ProofHistoryVoteResponse Result 내 필드 타입을 long -> int로 변경

+++ module-api/src/main/java/com/trybe/moduleapi/proof/dto/response/ProofHistoryVoteResponse.java
@@ -6,8 +6,8 @@ public class ProofHistoryVoteResponse {
     ) { }
 
     public record Result(
-            long approvedCount,
-            long disapprovedCount,
-            long nonParticipatedCount
+            int approvedCount,
+            int disapprovedCount,
+            int nonParticipatedCount
     ) { }
 }
